Altered design for FileHandler to return a FileBrowser and for...
the FileBrowser to contain all of the logic, removed from FileHandler
