+++ org.eclipse.ice.commands/src/main/java/org/eclipse/ice/commands/IFileHandler.java
@@ -12,7 +12,6 @@
 package org.eclipse.ice.commands;
 
 import java.io.IOException;
-import java.util.ArrayList;
 
 /**
  * This interface defines and lays out the design for the FileHandler structures
@@ -68,23 +67,10 @@ public interface IFileHandler {
 	public abstract void checkExistence(final String source, final String destination) throws IOException;
 
 	/**
-	 * This method returns the list of files and their relative paths to the passed
-	 * argument topDirectory. The argument should be the top-most directory that one wants
-	 * to search through.
+	 * Method that returns a FileBrowser instance
 	 * 
-	 * @param topDirectory
 	 * @return
 	 */
-	public abstract ArrayList<String> listFiles(final String topDirectory);
-
-	/**
-	 * This method returns the list of directories and their relative paths to the passed
-	 * argument topDirectory. The argument should be the top-most directory that one wants
-	 * to search through.
-	 * 
-	 * @param topDirectory
-	 * @return
-	 */
-	public abstract ArrayList<String> listDirectories(final String topDirectory);
+	public abstract FileBrowser getFileBrowser();
 
 }
