+++ org.eclipse.ice.commands/src/test/java/org/eclipse/ice/tests/commands/LocalFileHandlerTest.java
@@ -13,13 +13,7 @@ package org.eclipse.ice.tests.commands;
 
 import java.io.File;
 import java.io.IOException;
-import java.nio.file.DirectoryNotEmptyException;
 import java.nio.file.FileSystems;
-import java.nio.file.Files;
-import java.nio.file.NoSuchFileException;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.ArrayList;
 
 import org.eclipse.ice.commands.CommandStatus;
 import org.eclipse.ice.commands.Connection;
@@ -180,142 +174,6 @@ public class LocalFileHandlerTest {
 		fileCreator.deleteLocalSource();
 
 		System.out.println("Finished testing testExists()");
-	}
-
-	/**
-	 * Function to execute the local file browsing and local directory browsing
-	 * test. We call one main function so that a file structur can be created at the
-	 * beginning, accessed by both "subtests", and then deleted at the end.
-	 * 
-	 * @throws IOException
-	 */
-	@Test
-	public void testLocalBrowsing() throws IOException {
-		String topDirectory = "fileBrowsingDir";
-
-		Path topPath = createLocalFileStructure(topDirectory);
-
-		testLocalFileBrowsing(topPath);
-
-		testLocalDirectoryBrowsing(topPath);
-
-		deleteLocalFileStructure(topPath);
-	}
-
-	/**
-	 * Test for file browsing on local system
-	 * 
-	 * @throws IOException
-	 */
-	public void testLocalFileBrowsing(Path topPath) throws IOException {
-		LocalFileHandler handler = new LocalFileHandler();
-
-		ArrayList<String> files = handler.listFiles(topPath.toString());
-
-		// four files were created
-		assert (files.size() == 4);
-
-		// Let's assert that each file we created is in the list
-		for (int i = 0; i < files.size(); i++) {
-			String file = files.get(i);
-			// assert that the file is actually there
-			assert(handler.exists(file));
-			Path path = Paths.get(file);
-			
-			// assert that it is a file
-			assert (Files.isRegularFile(path));
-			
-		}
-
-	}
-
-	/**
-	 * Test for directory browsing on local system
-	 */
-	public void testLocalDirectoryBrowsing(Path topPath) throws IOException {
-		LocalFileHandler handler = new LocalFileHandler();
-
-		ArrayList<String> files = handler.listDirectories(topPath.toString());
-		for (int i = 0; i < files.size(); i++) {
-			System.out.println(files.get(i));
-		}
-		// Four directories total, including the top directory
-		assert (files.size() == 4);
-
-		// Let's assert that each directory we created is in the list
-		for (int i = 0; i < files.size(); i++) {
-			String file = files.get(i);
-			// assert that it exists
-			assert(handler.exists(file));
-			
-			// assert that it is a directory
-			Path path = Paths.get(file);
-			assert(Files.isDirectory(path));
-		}
-		
-	}
-
-	/**
-	 * Create a local file structure for testing local file and directory browsing
-	 * 
-	 * @param topDir
-	 */
-	protected Path createLocalFileStructure(String topDir) throws IOException {
-		// Create a destination path name
-		Path destinationPath = null;
-		// Try to create the path
-		try {
-			destinationPath = Files.createTempDirectory(topDir);
-		} catch (IOException e) {
-			e.printStackTrace();
-		}
-		String fullPath = destinationPath.toString();
-		System.out.println("Created top directory at: " + fullPath);
-		String separator = "/";
-		if (System.getProperty("os.name").toLowerCase().contains("win"))
-			separator = "\\";
-		Files.createDirectory(Paths.get(fullPath + separator + "dir1"));
-		Files.createDirectory(Paths.get(fullPath + separator + "dir2"));
-		Files.createDirectory(Paths.get(fullPath + separator + "dir3"));
-
-		Files.createFile(Paths.get(fullPath + separator + "filename.txt"));
-		Files.createFile(Paths.get(fullPath + separator + "dir1" + separator + "otherfile.txt"));
-		Files.createFile(Paths.get(fullPath + separator + "dir2" + separator + "newfile.txt"));
-		Files.createFile(Paths.get(fullPath + separator + "dir2" + separator + "newotherfile.txt"));
-
-		return destinationPath;
-	}
-
-	/**
-	 * Delete the local file structure for testing local file and directory browsing
-	 * 
-	 * @param topDir
-	 */
-	protected void deleteLocalFileStructure(Path topPath) {
-		// Delete the destination path that was created
-		System.out.println("Deleting " + topPath.toString());
-		try {
-			Files.deleteIfExists(topPath);
-		} catch (NoSuchFileException e) {
-			System.err.format("%s: no such" + " file or directory%n", topPath);
-			e.printStackTrace();
-		} catch (DirectoryNotEmptyException e) {
-
-			// If the directory is not empty, that is because it was a move command
-			// and the moved file is in there. So delete the file first and then
-			// delete the directory
-			File localDestinationFile = new File(topPath.toString());
-			boolean deleted = deleteLocalDirectory(localDestinationFile);
-
-			// Something went wrong and couldn't be deleted
-			if (!deleted) {
-				System.err.println(e);
-				e.printStackTrace();
-			}
-		} catch (IOException e) {
-			System.err.println(e);
-			e.printStackTrace();
-		}
 	}
 
 	/**
