+++ org.eclipse.ice.commands/src/test/java/org/eclipse/ice/tests/commands/RemoteFileHandlerTest.java
@@ -19,7 +19,6 @@ import java.nio.file.Files;
 import java.nio.file.NoSuchFileException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.ArrayList;
 import java.util.Collection;
 
 import org.eclipse.ice.commands.CommandStatus;
@@ -298,168 +297,6 @@ public class RemoteFileHandlerTest {
 
 	}
 
-	/**
-	 * Function to execute the remote file browsing and remote directory browsing
-	 * test. We call one main function so that a file structure can be created at
-	 * the beginning of the test, accessed by both "subtests", and then deleted at
-	 * the end
-	 * 
-	 * @throws Exception
-	 */
-	@Test
-	public void testRemoteBrowsing() throws Exception {
-		// We'll create our own code here to create the file structure rather than
-		// using the already developed functions, since we need multiple files and
-		// directories
-		String topDirectory = "/tmp/fileBrowsingDir/";
-
-		// put this in a try and finally so that the remote file structure
-		// always gets deleted. Otherwise the next test run might fail when it
-		// tries to build the remote file structure and see that it already
-		// exists
-		try {
-			createRemoteFileStructure(topDirectory);
-
-			testRemoteFileBrowsing(topDirectory);
-
-			testRemoteDirectoryBrowsing(topDirectory);
-		} finally {
-			deleteRemoteFileStructure(topDirectory);
-		}
-	}
-
-	/**
-	 * Test for file browsing on remote system
-	 * 
-	 * @throws IOException
-	 * @throws SftpException
-	 */
-	public void testRemoteFileBrowsing(String topDirectory) throws IOException, SftpException {
-
-		RemoteFileHandler handler = new RemoteFileHandler();
-		handler.setConnectionConfiguration(fileTransferConn.getConfiguration());
-
-		ArrayList<String> files = handler.listFiles(topDirectory);
-
-		// files should only be 4 entries since there are only 4 files in the tree
-		// structure we created
-		assert (files.size() == 4);
-
-		for (int i = 0; i < files.size(); i++) {
-			// Assert that the file exists
-			assert (handler.exists(files.get(i)));
-
-			ChannelSftp channel = fileTransferConn.getSftpChannel();
-			Collection<ChannelSftp.LsEntry> structure = channel.ls(files.get(i));
-			// This should just be a one file vector, since we are ls-ing the filename
-			for (ChannelSftp.LsEntry filename : structure) {
-				// Assert that the file is indeed a regular file
-				assert (filename.getAttrs().isReg());
-			}
-		}
-	}
-
-	/**
-	 * Test for directory browsing on remote system
-	 * 
-	 * @throws IOException
-	 * @throws SftpException
-	 */
-	public void testRemoteDirectoryBrowsing(String topDirectory) throws IOException, SftpException {
-
-		RemoteFileHandler handler = new RemoteFileHandler();
-		handler.setConnectionConfiguration(fileTransferConn.getConfiguration());
-
-		ArrayList<String> files = handler.listDirectories(topDirectory);
-
-		// directories should only be 3 entries since there are only 3 directories in
-		// the tree structure we created
-		assert (files.size() == 3);
-
-		// Use the default separator of the remote dummy system
-		String separator = "/";
-
-		for (int i = 0; i < files.size(); i++) {
-			assert (handler.exists(files.get(i)));
-			ChannelSftp channel = fileTransferConn.getSftpChannel();
-
-			// ls the previous directory so that we can look at the subdirectories
-			Collection<ChannelSftp.LsEntry> structure = channel
-					.ls(files.get(i).substring(0, files.get(i).lastIndexOf(separator)));
-			// Iterate through the ls
-			for (ChannelSftp.LsEntry filename : structure) {
-				// Check that the ls filename is actually in the directory list (and thus is a
-				// directory)
-				if (files.get(i).contains(filename.getFilename())) {
-					// assert that it is a directory
-					assert (filename.getAttrs().isDir());
-				}
-			}
-		}
-
-	}
-
-	/**
-	 * Function that creates a dummy remote file structure tree to test the file
-	 * browsing source code
-	 * 
-	 * @param - topDirectory - refers to top directory whose contents will hold the
-	 *          dummy directories/files
-	 * @throws SftpException
-	 */
-	protected void createRemoteFileStructure(String topDirectory) throws Exception, SftpException {
-
-		ChannelSftp sftpChannel = fileTransferConn.getSftpChannel();
-
-		// Check if the directory already exists
-		SftpATTRS attrs = null;
-		try {
-			attrs = sftpChannel.lstat(topDirectory);
-		} catch (Exception e) {
-			System.out.println("Remote directory not found, trying to make it");
-		}
-		if (attrs == null) {
-			// Remote directory doesn't exist, so make it
-			// Create a remote source directory
-			sftpChannel.mkdir(topDirectory);
-		}
-
-		// make another directory in top directory
-		sftpChannel.cd(topDirectory);
-		sftpChannel.mkdir("dir1");
-		sftpChannel.mkdir("dir2");
-		sftpChannel.mkdir("dir3");
-
-		// create a local file to put there
-		createLocalSource();
-		// Get the filename that was just created
-		String filename = theSource;
-		// put it in a few places in the directory structure
-		sftpChannel.put(filename, topDirectory);
-		sftpChannel.put(filename, topDirectory + "/dir1/");
-		sftpChannel.put(filename, topDirectory + "/dir3/");
-		sftpChannel.put(filename, topDirectory + "/dir3/newfile.txt");
-
-	}
-
-	/**
-	 * Function that deletes the remote file structure tree created for testing file
-	 * browsing
-	 * 
-	 * @param - topDirectory - refers to top directory whose contents hold the dummy
-	 *          directories/files
-	 * @throws JSchException
-	 * @throws SftpException
-	 */
-	protected void deleteRemoteFileStructure(String topDirectory) throws SftpException, JSchException {
-		// Connect the channel from the connection
-		ChannelSftp sftpChannel = fileTransferConn.getSftpChannel();
-
-		System.out.println("Deleting remote destination at : " + topDirectory);
-		// Recursively delete the directory and its contents
-		deleteRemoteDirectory(sftpChannel, topDirectory);
-
-	}
 
 	/**
 	 * Dummy function which makes the connection configuration for the dummy remote
